value-pairs: Introduce value_pairs_walk()
This new function is meant to be used when one wants to process the
result of value-pairs() in a structured manner. It will treat the keys
as if they were in dotted-notation, and whenever a new object starts,
it will call obj_start_func, whenever an object ends (a new one
starts, or we're winding down in the end), it calls obj_end_func. When
processing a single key, process_value_func will be called.

The obj_start_func and obj_end_func callbacks take six arguments: the
current key name, the prefix so far, a pointer to an area where
prefix-specific data can be stored, the previous element's name and
data on the stack, and the user_data passed to value_pairs_walk(). The
process_value_func receives the value too, see the VPWalk* typedefs in
value-pairs.h.

The prefix-specific data can be NULL, so care must be taken not to
dereference it in that case.

With this functionality, we will be able to construct structured data
out of the flat representation of value-pairs.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>