afmongodb: Use insert instead of upserts
With the recent changes to value pairs, we can now turn our internal
dotted notation into structures, so do that, and use inserts instead
of upserts!

This - at the moment - comes with a noticable performance cost, as
we'll do extra work on the syslog-ng side (which was previously done
on the server side). However, this paves the way for bulk inserts,
which will improve performance a lot. (And making the value pairs
walker more performant will also help, but that's independent of
mongodb.)

Signed-off-by: Gergely Nagy <algernon@balabit.hu>