system-source: Introduce a new system() generator.
This patch replaces the old confgen-based system() generator, that
relied on a script (generate-system-source.sh) to generate the block
system() expands into. While that was flexible and easy to modify, it
has had the disadvantage of running an external script at startup,
which is undesirable in many situations, including when using a strict
AppArmor profile.

However, a generated system() is useful in a lot of situations,
ranging from creating portable, system-agnostic configuration files to
supporting sysvinit and systemd with the same config.

For these reasons, this patch implements a built-in system()
generator, that does not rely on the generate-system-source.sh script
at all, but does everything itself, from C.

It is loaded by the default scl.conf, but can be loaded separately
with:

 @module system-source

It does not support auto-loading yet, because all the module does, is
register a block.

The system() generator this patch implements results in the exact same
expansion the old script generated. The old script-based solution is
also deprecated (and thus, removed), in order to not duplicate code:
we only wish to maintain one implementation.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>