cfg-lexer: Fix glob-based include order
When using glob-based includes, matching files were immediately
included, which resulted in the include stack going bonkers. The
proper way to do glob-based includes is to do the same as in the
directory-based case: add the files to the current include stack
level's file list.

This patch does just that, thereby making glob-based includes work as
one would expect them to work.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>