doc: Move old ChangeLogs to doc/
The old ChangeLogs (especially ChangeLog) were looking stale in the
root directory, and ChangeLog was overwritten at make dist
anyway. Since they're historic, move them to doc/: they'll still be
available in the git repo, but they will not be included in the
tarball.

The top-level makefile will continue to re-generate ChangeLog
appropriately, so that the built tarballs will continue to have an
up-to-date changelog.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>