afsocket: In case of unix-dgram sources, include the filename in stats
The file source includes the filename of the source in the stats, but
the unix-dgram source does not do the same, which makes it harder to
identify which unix-dgram stat belongs to which source.

With a little change to afsocket_sc_stats_instance(), we can fix that,
by teaching this function to use self->owner->bind_addr if
self->peer_addr is unset.

Reported-by: Evan Rempel <erempel@uvic.ca>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>