simplify *Options initialization pattern (LogReaderOptions, LogWriterOptions, etc)
The _init() method for various *Options structures were needlessly
complicated with an ugly warning on top of each and every such function.
After giving some thought to why it needs to be so complicated a much
easier pattern could be created, eliminating most of the complexity.

Earlier *_options_init() methods were prepared to be called several times
without a pairing _options_destroy(), e.g. if any memory was allocated
by init, a repeated init should have preserved the allocation (or at least
free the one the previous invocation allocated). This pattern caused
these init methods to actually save a number of variables locally
before commencing on the actual initialization and then restoring
them afterwards.

After giving some thought, this actually addressed an unused use-case:
if the init method is called for the same Options structure with
a changed GlobalConfig instance. In this case init would've inherited
the changed values from GlobalConfig.

However this never happens in reality, options structures may indeed be
initialized multiple times, however in this case the GlobalConfig remains
the same.

Removing this assumption simplifies the code a lot and makes it much more
readable, and finally stops spreading that terrifying comment on the top
as the pattern got implemented in a number of places.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>