afmongodb: Include libmongo-client/configure.gnu in the tarball
In order to install cleanly with an internal libmongo-client, we need
that library to be compiled in statically, which is what the
configure.gnu script (installed via autogen.sh from git) enforces.

However, since configure.gnu is not part of libmongo-client, its make
dist will not include it, so its parent must do that. For this reason,
and similarly to how the same file in lib/ivykis is handled, include
libmongo-client/configure.gnu in the EXTRA_DIST of
afmongodb/Makefile.am.

Reported-by: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>