[affile] Affile source driver save the own reader to the persist config because
the reader store the actual window size. In certain cases dropping messages can occured
although the flow control is turned on after reload config
The root cause was, that the file source driver dropped the reader and create a new one
under reload process, but the actual window_size is also dropped with reader.
The solution is that store the reader into the persist config and the source driver
try it to fetch from the persist config

Signed-off-by: Juhasz Viktor <jviktor@balabit.hu>

Conflicts:

	modules/affile/affile.c