debian: build-depends update & module dependency tightening
We need python and automake >= 1:1.11.1 to build, and the modules
should depend on syslog-ng-core (= ${binary:Version}). The makeshlibs
magic was supposed to help with that, but that only works if the -dev
package is already installed.

There's probably a way to do this better, but for now, this will do.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>