travis: Significant Travis CI control file updates
To fix issues with uninstallable packages, and to follow the official
recommendation, run apt-get update as part of the testing process (and
also install libesmtp-dev). Also, to make testing more reliable,
install into $HOME, so that we do not need sudo for running the tests.

Furthermore, restrict the tested branches to a subset: master,
upstream mirrors, the sandbox and merge-queue branches. While there,
also enable IRC notifications.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>