[dbparser] Implemented EMAIL parser
The EMAIL parser will match address. The parameter is a set of
characters to strip from the front and the end of the e-mail address
which allows for easy matching of <email@some.host> enclosed
addresses. Characters that a valid for a hostname can not be stripped
from the end of the hostname. This includes a trailing period if
present.

Reported-by: Evan Rempel <erempel@uvic.ca>
Signed-off-by: Evan Rempel <erempel@uvic.ca>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>