tfgeoip: $(geoip) template function
This implements a $(geoip) template function using libgeoip, based on
a patch from Csaba Major. The function can be used to look up the
2-letter country code of any IPv4 address or host passed to the function.

Currently only the 2-letter codes are supported, and only from the
default database, and only for IPv4 addresses (or hosts). Adding
further functionality can be done on demand, however.

Signed-off-by: Csaba Major <csaba.major@balabit.com>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>