cfg-lexer: Handle no-magic includes better, without hiding errors
When the @include pragma was converted to use glob(3), a couple of
problematic cases stopped producing errors: if we had an @include
without any magic, it still went through glob(3), which was set up to
ignore errors (so that @include "/etc/syslog-ng/conf.d/" would work,
even when the directory didn't exist).

This hid errors, and made it possible even for explicit, glob-less
@includes to fail silently, making debugging next to impossible.

To fix this, glob() is now called with GLOB_NOMAGIC, so if the pattern
does not contain any magic, it is returned anyway, whether it exists
or not: and then, it will eventually be tried by the ordinary include
mechanism, and produce a failure there, as it is appropriate.

This way, @include "/foo/*" continues to work, and @include
"/nothing.conf" will fail, as it should: errors are only ignored for
inclusions that contain magic characters.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>