debian: Route internal messages to /var/log/syslog-ng.log
To avoid running into a self-DoS when debugging is enabled along with
heavy filtering, move internal messages to /var/log/syslog-ng.log by
default, without filters.

This is done by pulling the internal source out of s_src, and making a
new source for it, and connecting that to the new destination.

Reported-by: Anton <koldaevav@gmail.com>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>