messages: only allow a single level of recursions
The code as originally written has permitted a configurable number
of recursions to detect the positive feedback loop of internal messages.
This feedback loop may happen when an internal() message triggers
another internal() which in turn again and so on, effectively eating
all the available system memory.

Since we've used a single level of recursions anyway, this patch simplifies
the related code and frees up 6 bits in the LogMessage structure.

It also adds some more information into the suppression message, so it can
be easier to understand.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>