Major copyright & license fixups
Update all the copyright years in pretty much any file touched in the
past 2-2.5 years, and clarify the licensing situation:

lib/ is LGPL2.1+, modules/ is GPL2+, both with an OpenSSL
exception. Updated the top-level COPYING file to reflect the changed
paths too, and LGPL.txt was replaced with the text of LGPL 2.1
(instead of LGPL 3, when the sources were under 2.1+).

Signed-off-by: Gergely Nagy <algernon@balabit.hu>