dbparser: support inheriting properties in a corellation action
This patch implements an inherit-properties attribute for message type
actions is patterndb correlation. In case this is set to TRUE, the
triggering log message is cloned and all name-value pairs (including
tags) are automagically copied to the emitted message. The values set
in the values XML subtree overwrite the original values. In case it is
set to FALSE or unspecified, a new message is generated and only the
values set manually are added to it.

   <actions>
     <action>
       <message inherit-properties='TRUE'>
         <values>
           <value name="PROGRAM">my-program-name</value>
           <value name="MSG">Hello new message!</value>
           <value name="contextid">${session-id}@2</value>
         </values>
       </message>
     </action>
   </actions>

Signed-off-by: Balint Kovacs <blint@blint.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>