value-pairs: Support rekey outside of key()
There are situations where one would like to apply transformations to
keys selected by the scope, without having to re-list the selected
keys in the config (such as default transformations for mongodb and
format-json). For this to work, we need to be able to use rekey
outside of key(), so that the pattern given to rekey will not be
explicitly added to the set.

In other words, this allows us to write transformations that work on
the already assembled set, without adding new elements to it.

For the command-line variant, this is done with the new --rekey
option, which is exactly like --key, except it does not add the
pattern to the set.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>