Use value-pairs transformations in mongodb and format-json
For both mongodb and json, leading dots are not allowed on key
names. The mongodb destination had a guard against this, and replaced
leading dots with an underscore, format-json did not.

With this patch, both will forcibly add a value-pair key
transformation that replaces the first leading dot. This covers the
most common cases, but makes the mongodb destination vulnerable to
keys like ".foo..bar". However, it is not the purpose of the driver to
guard against corner cases.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>