dbparser: Make the context length available to action conditions
It was quite unintuitive to determine if a previous message exists in
the corellation context as the basis of the decision whether an action
should be run or not. This patch implements a $(context-length)
template function that returns the number of messages within a
corellation context, and can be used in the condition of an action.

This way it is easy to determine if a context contains sufficient
amount of messages for the event to be complete and run a different
action if it is not, e.g. if 2 messages are needed for an event, this
should emit an event message if there are enough messages and an error
message if not.

 <actions>
  <action condition='"$(context-length)" >= "2"'>
   <message>
    <values>
     <value name="PROGRAM">event</value>
     <value name="MESSAGE">Hello event!</value>
    </values>
   </message>
  </action>
  <action condition='"$(context-length)" < "2"'>
   <message>
    <values>
     <value name="PROGRAM">error</value>
     <value name="MESSAGE">Hello error!</value>
    </values>
   </message>
  </action>
 </actions>

(Based on a patch from Balint Kovacs <blint@blint.hu>, and
suggestsions from Balazs Scheidler <bazsi@balabit.hu>)

Requested-by: Balint Kovacs <blint@blint.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>