[crypto] New ivykis on windows using one thread for each handle waiting
it means, that a lot of thread started and terminated while the syslog-ng running.
The openssl is initialized, so in every thread the ssl_thread_id will be called.
But the g_thread_self allocate memory if the thread doesn't started as g_thread.
In case of many threads it cause a huge memleak.

Signed-off-by: Juhasz Viktor <jviktor@balabit.hu>