[affile_source] affile.c: The "/dev/klog" cannot pollable under FreeBSD 9.1
so we set the follow_freq to 1 sec and we drop a warning message.

Signed-off-by: Karsai Csaba <cskarsai@balabit.hu>

+

From: Gergely Nagy <algernon@balabit.hu>
Date: Fri, 5 Oct 2012 12:16:57 +0200
Subject: [PATCH] logreader: When following a file, treat non-regular files
 specially

When following a file with follow-freq(), and it is not a regular
file, but something for which stat(2) always returns 0 size (such as
the /dev/klog character device on FreeBSD),
log_reader_io_follow_file() never went into the process input branch,
because the current position (0) was always at EOF (also 0).

To break out of this unfortunate situation, if the file to read from
is not a regular file, go into the process input loop anyway: at
worst, it will do an expensive noop.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>