afsocket: Add an ip-protocol() option for the network driver
To make it easy to set up IPv6 sources or destinations, the network()
driver should have an option to select the IP protocol. This patch
does just that, in the form of an ip-protocol() option, which allows
one to select either IPv4 or IPv6.

In the end, this allows us to have a source as follows:

source s_v6_syslog {
  network(port(12345)
          transport(framed)
          ip-protocol(6));
};

The option is available for both the source and destination version of
network().

Signed-off-by: Gergely Nagy <algernon@balabit.hu>