afamqp: Include rabbitmq-c/configure.gnu in the tarball
In order to install cleanly with an internal rabbitmq-c, we need that
library to be compiled in statically, which is what the configure.gnu
script (installed via autogen.sh from git) enforces.

However, since configure.gnu is not part of rabbitmq-c, its make dist
will not include it, so its parent must do that. For this reason, and
similarly to how the same file is handled elsewhere, include
rabbitmq-c/configure.gnu in the EXTRA_DIST of afamqp/Makefile.am.

Reported-by: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>