cfg-lexer: Support systems without GLOB_NOMAGIC
GLOB_NOMAGIC is a GNU-ism, and not present on all systems syslog-ng
supports, so for those system, we need a workaround. We can't use
GLOB_NOCHECK, as that would make us have widly different behaviour in
some cases (when there really is a pattern in the @include'd filename).

So, on platforms that do not have GLOB_NOMAGIC, we implement the same
behaviour on our own: if glob() failed with GLOB_NOMATCH, we check if
the filename is a pattern, and if not, add the pattern to the include
list as-is, otherwise silently ignore it.

This means that @include "/etc/syslog-ng/conf.d/*.conf" will work on
these platforms the same way as it does on GNU-enabled ones.

Reported-by: Ben Lentz <blentz@cswg.com>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>