Make it possible to link crypto stuff directly into libsyslog-ng
There's been a long desire to be able to link libsyslog-ng-crypto
directly into libsyslog-ng, if so desired. To accomplish that, we
introduce a --with-embedded-crypto option to configure, which in turn
produces a WITH_EMBEDDED_CRYPTO automake conditional.

If that conditional is set, then the sources of libsyslog-ng-crypto
get appended to the libsyslog-ng sources, and -crypto is not
built. All users of the -crypto module (afsocket) have been updated to
not try and link in libsyslog-ng-crypto if WITH_EMBEDDED_CRYPTO is
set.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>