Make it possible to link crypto stuff directly into libsyslog-ng
There's been a long desire to be able to link libsyslog-ng-crypto
directly into libsyslog-ng, if so desired. To accomplish that, we
introduce a --with-embedded-crypto option to configure, which in turn
produces a WITH_EMBEDDED_CRYPTO automake conditional.

If that conditional is set, then the sources of libsyslog-ng-crypto
get appended to the libsyslog-ng sources, and -crypto is not built. In
the non-embedded case CRYPTO_LIBS will be set to
${top_builddir}/lib/libsyslog-ng-crypto.la, otherwise it will be
empty. All users of the -crypto module (afsocket, dbparser &
cryptofuns) have been updated to use $(CRYPTO_LIBS) instead of
hard-coding libsyslog-ng-crypto.la.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>