syslog-ng-ctl: Fix the linking with modern libtool
The Makefile.am for syslog-ng-ctl relied on lib/gsocket.c and
lib/gsockaddr.c getting compiled to gsocket.lo and gsockaddr.lo,
respectively. This is not always the case, sometimes the library name
gets prepended.

To work around this, add the two source files to syslog_ng_ctl_SOURCES
instead. This way, they'll get compiled again, but the result will
always work, and not rely on fragile convention.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>