debian: Use /dev/ttyva on kFreeBSD as the d_console_all target
On kFreeBSD, /dev/ttyva should be used instead of /dev/tty10. To keep
the main config similar between all architectures, split the tty10
definition out of it, and move it to an include file instead.

At install time, we'll decide which definition to include, based on
what architecture we're building for.

Closes: Debian#697042
Reported-by: Christoph Egger <christoph@debian.org>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>