affile: Restore fsync & flush-lines functionality
Back in May, in 101cb379, which moved fsync() from logtransport to
logproto, a new argument for log_proto_file_writer_new() was
introduced, but when updating the call sites, the flush_lines and
fsync arguments were mixed up.

This corrects that, restoring both flush-lines() and fsync()
functionality for the file destination.

Reported-by: Evan Rempel <erempel@uvic.ca>
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>