filter: add filter invocation functions that automatically make the LogMessage writable
In case a filter is not read-only, we have to make sure that our LogMessage
instance is writable. Introduce a new set of function that does
exactly that.

The old ones still exist as in some cases filter invocation is not
on a log processing path (patterndb actions, etc), in which case we
still have to be able to invoke them.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>