LogProtoBufferedServerState: deprecate buffer_cached_eol member
There are multiple reasons it was a bad idea to stored buffer_cached_eol
in the state structure:
  - the buffer_cached_eol member belongs to LogProtoTextServer in the first
    place
  - the cached_eol is just a cache, it could have been recalculated
    anyway when the state is reloaded, the least concern we have at restart
    time is that we need to relocate the closing EOL mark in our buffer
  - for multi-line support more state would be required.

Since this affects the persist file format the patch needs to work with
old persist files properly. For that reason the member is not completely
removed, but a dummy entry remains. We always set that member to 0, whenever
we load or upgrade a persist entry.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>