LogMultiplexer: propagate branch properties to the head of the branch
Since the compilation process doesn't always have all the context to
propagate branch-related LogPipe flags back to the head element
of the LogMultiplexer branches, do that in log_multiplexer_init().

We basically pick up "final" and "fallback" flags from Pipes on a branch
regardless if they have been set on the branch head or on subsequent
elements.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>