LogProtoTextServer: implement accumulate_line and refactor
This patch introduces a new accumulate_line() method that allows the use of
continuation lines.

Testcases were split to give a better indication of what went wrong in the
code, and a number of new testcases were added.

Functions in the implementation were broken down to smaller ones, to make
the code easier to follow.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>