plugin-types.h: new header to define LL_CONTEXT_* macros
This role of this header is to define the LL_CONTEXT_* macros used by
various plugin implementations. For now it simply includes the cfg-grammar.h,
in the future it could remove the namespace poisoning that this causes by
using a different means.

The reason this became a dedicated header is that it is sometimes difficult
to grasp why a given code requires code generated by bison from the main
configuration file parser.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>