upgradetool: This patch implements the upgradetool.
upgradetool does the upgrade windows agent config from version: 4.0.x to 5.0.1
libxml2 is added to the syslog-ng dependencies, because the agent-config module
will use this library to validate and upgrade the xml config file.

GPO specific functionality comes from the previous upgradetool project,
which was implemented in VC++.

The upgrade tool generate a log html file which has formatted output about,
what was modified and why.

Signed-off-by: Juhasz Viktor <jviktor@balabit.hu>