affile: using wildcards in file sources could lead to message loss
If simple filename is used in filesource(ie. doesn't contain wildcard)
syslog-ng set the related file_monitor to NULL(it's OK because for
one file we don't need file monitor object and don't have to deal with
any queue containing file names). In case of NC_FILE_SKIP, file_monitor
is checked. If it is NULL, the related log reader is restarted,
thus when the file become readable, syslog-ng will able to read it.
When wildcard is used in name of the file source, syslog-ng handles a
list of file names(queue) and a file monitoring object.
When the queue becomes empty or it has an element which cannot be opened
(because its renamed, deleted, etc...), syslog-ng should restart the
related log reader, otherwise it won't react any changes to that file.

References: v1 D-05647
References: bugzilla #24331