resolve_sockaddr: IP address truncated in some cases
When hostname cannot be determined by getnameinfo(), the IP address is returned
by default. When use_fqdn is unset(set to false, by default) that IP address
will be the input of hostname truncating(-> IP address will be truncated).

It is possible to modify behaviour of getnameinfo() via flags: if NI_NAMEREQD
is set, getnameinfo() will return an error instead of IP address when hostname
cannot be determined.