build: use non-recursive automake
This is a huge change, one that cannot be sensibly broken up into
smaller pieces, so please bear with me! The purpose of the complete
rewrite of the build system is to get rid of recursive make calls, and
instead, use a single, non-recursive Makefile built up from multiple
included parts, that has all the dependency information, therefore
providing a much more robust dependency tracking, which is also
significantly faster.

For the sake of convenience, the Mk/populater-makefiles.sh script will
add wrapper Makefiles to every directory that has a Makefile.am (unless
it belongs to a third party library), allowing one to restrict 'make
all' and 'make check' to a particular directory. The 'make check'
provided by this convenience Makefile does not recurse into
subdirectories, however.

The system also provides shortcuts to build various parts of syslog-ng:
 - make syslog-ng
 - make mod-sql mod-mongodb mod-socket
 - etc

Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>