afsocket: Avoid a use-after-free when connection is alive but driver is freed.
We register a callback for certificate hostname verification in the destination driver.
This callback can be called anytime during the TLS connection lifetime, for
example, when renegotiation happens, or during deferred handshake. We should
ref the destination driver structure to make sure, that the driver isn't freed,
when the verification happens. The connection is persistent during reload,
so the old driver should be reachable after reload too.