afsocket: Support a smooth upgrade path to recent systemd
When the journal was introduced in systemd, it started to pass
/run/systemd/journal/syslog to syslogds (it was passing down /dev/log
before). Since we verify that the FD to open matches the expected
filename, we didn't pick up the fd due to the filename mismatch.

This can be fixed on the system() source level, where it can decide
which to generate into the config. However, for setups that do not use
system() for one reason or the other, this does not work.

For these systems that do not use the system() source, the best
solution is to try and replace /dev/log with
/run/systemd/journal/syslog at run-time. This patch does just that,
and ties it to config versions 3.4 or earlier.

This means that post-3.4, the automatic replacement will stop working,
and users will get the generic failed to acquire socket error message
instead.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>