LogReader: introduce the LogReaderPollEvents interface
This is an intermediate step that introduces a LogReader specific private
class to perform I/O polling.  In the subsequent patches this will become
more general so it can be applied by LogWriter instances as well.  This
patch changes the LogReader logic to instantiate the appropriate class at a
single location and rely on LogReaderPollEvents descendants to perform the
work it needs but with a uniform interface.

This patch introduces the abstract base class (LogReaderPollEvents) and two
descendants factored out of LogReader, LogReaderPollFileChanges and
LogReaderPollFd.  The first performing the follow-freq() style time based
polling, the second merely polling the specified fd using poll(2).

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>