scl: On HPUX, we should use log_fetch_limit(1000) for /dev/klog
Syslog-ng stucks on reading /dev/klog during boot. The problem lies around log_reader_fetch_log.
There is the may_read variable which disables calling read() in log_proto_fetch implementations
, only fetching from buffered data is allowed. We set this may_read to false when LPBS_NOMREAD
is set, and it is set, because follow_freq is 0. But every time we reach fetch_limit, we return
from log_reader_fetch_log, and next time we enter it, may_read is TRUE again, so
log_proto_buffered_server_fetch will call read() again, and we will block, because /dev/klog on
hpux is blocking, event when O_NONBLOCK is set. A crude workaround is to increase log-fetch-limit
in /dev/klog source to 1000, so we can read all the buffer in one log_reader_fetch_log call.