PollEvents: implement PollEvents abstract class
The new PollEvents class is factored out from LogReader as a generic
interface to implement event polling both in LogReader and LogWriter.

The plan is to have a generic PollFdEvents class to poll for fd I/O events,
and a separate class for file-change polling when using follow-freq().

Other polling methods may become available in the future.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>