LogReader: drop constructor arguments in favour of log_reader_reopen
Instead of passing the LogProtoServer and PollEvents instance in the
constructor, pass those using log_reader_reopen() once the constructor
has finished. This makes LogWriter and LogReader a bit more similar and
provides a single path for passing this data.

This patch also cleans up the previously unused log_reader_reopen()
interface and removes most of the arguments, which should've been
passed to more specialized methods.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>