LogProtoTextServer: make it possible to drop the end of a line in accumulate_line()
This patch introduces LPT_CONSUME_PARTIALLY() macro that piggybacks
the number of characters to be dropped from the end of the line. This
is the processed by LogProtoTextServer right before returning the line.

Note that it is not possible to drop arbitrary line ends, it only works
at the end of the line, as characters are not moved in the input buffer in
any way.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>