loggen: fix division by zero
If the file given to the loggen with the read-file command line
parameter contains only one line that is in invalid format and
the dont-parse option is missing then the program exits from
the inner loop in the gen_messages function while the count
variable is still zero, and division by zero occurs.
From now in this case the raw_message_length global variable is
simply set to zero.

Signed-off-by: Daniel Gados <gdani@balabit.hu>