thread local storage: surround the struct with cookies
So TLS can be found while analyzing a core file in debugger, regardless of
which TLS implementation is in effect.

In case of the pthread_*() implementation, it's a little hairy to access
the internals of the __tls_variables struct during initialization, since
when the function is being compiled, the definition of the struct is
not created yet. Actually, that's why the size argument is required.
Since putting the cookies into the struct requires it to be defined when
the function doing the work is compiled, I needed to defer compilation of
the helper functions until the struct definition is generated by generating
the function from a macro. With this, the size argument became useless.

Conflicts:

	lib/tls-support.h