filemonitor: Teach resolve_to_absolute_path() to handle absolute paths correctly
resolve_to_absolute_path() handled the case of absolute symlinks
badly: it prepended the basedir anyway, whether the symlink is
absolute or relative. This lead to unavailable paths if a customer had
absolute directory symlinks for example.

This patch applies a quick band-aid: if the link_target begins with /,
we'll use it as-is, without prepending the basedir, otherwise we
proceed as before.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>