log_template_compile: function refactorized
This patch adds a LogTemplateCompiler that encapsulates the template
compilation process that is independent from actually formatting (or
evaluating) the template.  This reduces the number of parameters to parsing
functions and makes the code easier to follow.

The next possible step should be, that split the templates.c
file into more files.

This patch based on the log-template-refactor branch

Signed-off-by: Juhasz Viktor <jviktor@balabit.hu>