syslogformat: fix bad pointer cast that leads to timestamp overwriting (fixies: D-07734)
In the log_msg_parse_date function the variable 'now' is of type GTimeVal.
Its member tv_sec is of type long, and its address was cast to a time_t
pointer.
On Windows 64 bit the size of time_t is 64 bits (by default without
defining the _USE_32BIT_TIME_T macro), but the size of the long type
remains 32 bits. This led to invalid pointer casting and bad argument
passing to the cached_localtime function.

Signed-off-by: Daniel Gados <gdani@balabit.hu>